requirements.txt
blinker==1.6.2
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
Flask==2.3.2
fonttools==4.40.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
numpy==1.25.0
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
pyparsing==3.1.0
pytesseract==0.3.10
python-dateutil==2.8.2
six==1.16.0
tesseract==0.1.3
Werkzeug==2.3.6
